pyproject.toml
[tool.poetry]
name = "passwordmanager"
version = "0.1.0"
description = "A password manager with a Tkinter GUI and SQLCipher encryption"
authors = ["Travis Gooden"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
attrs = "22.1.0"
autopep8 = "2.0.0"
cachecontrol = "0.12.11"
cachy = "0.3.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
cleo = "1.0.0a5"
crashtest = "0.3.1"
cryptography = "38.0.4"
dependency-injector = "4.40.0"
distlib = "0.3.6"
dulwich = "0.20.50"
filelock = "3.8.2"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "5.1.0"
jaraco-classes = "3.2.3"
jeepney = "0.8.0"
jsonschema = "4.17.1"
keyring = "23.11.0"
lockfile = "0.12.2"
more-itertools = "9.0.0"
msgpack = "1.0.4"
packaging = "22.0"
peewee = "3.15.4"
pexpect = "4.8.0"
pillow = "9.3.0"
pkginfo = "1.9.2"
platformdirs = "2.6.0"
poetry = "1.2.2"
poetry-core = "1.3.2"
poetry-plugin-export = "1.2.0"
ptyprocess = "0.7.0"
pycodestyle = "2.10.0"
pycparser = "2.21"
pylev = "1.4.0"
pyrsistent = "0.19.2"
requests = "2.28.1"
requests-toolbelt = "0.9.1"
secretstorage = "3.3.3"
shellingham = "1.5.0"
six = "1.16.0"
tomli = "2.0.1"
tomlkit = "0.11.6"
urllib3 = "1.26.13"
validator-collection = "1.5.0"
virtualenv = "20.17.1"
webencodings = "0.5.1"
zipp = "3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
